"Prepare baking" button assigns correct material
The button now also assigns surfaces to the correct material type. Surfaces that are in the lightmapped vertex group will automatically get a material that bakes to the lightmap bake image and non lightmapped surfaces will automatically get a material that bakes to the vertmap.

Changing light type assignments doesnt require patching and reimporting bsps anymore with that change.